fix: set tanstack config to use base config without import plugin
Tanstack seems to handle import rules within their own eslint config. Letting import be included was causing a
build error, however the import plugin is still needed for other workspaces that use the base config. So
solution for now is to just split the configs, as done in 5f53a67c, and use the no import config.