refactor: rename @phyt/api from generic to clear @phyt/hono-gate
Sometimes beneficial to be generic (obfuscate function and tech in favor
of purpose). However, private repo begets clear naming and task
responsiblities. This package imports all core functionality and only
really serves as an api gateway layer. As for why the 'hono' --
possibility for other server stacks or impl. in the future, again
clarity.